+++ wuddle-gui/src/gif-decode.js
@@ -1,0 +1,5 @@
+/**
+ * Minimal gifuct-js re-export for bundling into vendor/.
+ */
+import { parseGIF, decompressFrames } from "gifuct-js";
+export { parseGIF, decompressFrames };